[feat] add blog, footer, newsletter components
[style] made the styles for the home container to better style the
learning recommendations

Note: I had to add the .env.sample file to the repo because I added a new
variable to the .env file. I also had to add the client.js file to the
repo so as to allow connection to the Sanity API for the blog posts.

The styles for the blog components still need to be worked on and made
better especially in responsiveness for the tablet view (1100px and
below)

[feat] Added new font variables to index.css file.

This variables can be used in any css (or scss) file in the project.

I also added new images to be used by the footer to the assets folder.

new packages included in the package.json file are:
- @sanity/client